requirements.txt
apache-airflow[postgres]
apache-airflow[s3fs]
pandas
apache-airflow-providers-postgres
pyarrow